+++ src/portable/wirestate-react-signals.ts
@@ -1,5 +1,4 @@
 export * from "@wirestate/core";
 
 export * from "../wirestate-react";
-export * from "../wirestate-react/test-utils";
 export * from "../wirestate-react-signals";