+++ src/wirestate-react/context/use-scope.test.tsx
@@ -1,7 +1,7 @@
 import { render } from "@testing-library/react";
 import { Container, WireScope, createContainer } from "@wirestate/core";
 
-import { withContainerProvider } from "../test-utils/with-container-provider";
+import { ContainerProvider } from "../provision/container-provider";
 import { Optional } from "../types/general";
 
 import { useScope } from "./use-scope";
@@ -17,7 +17,11 @@ describe("useScope", () => {
       return null;
     }
 
-    render(withContainerProvider(<TestComponent />, container));
+    render(
+      <ContainerProvider container={container}>
+        <TestComponent />
+      </ContainerProvider>
+    );
 
     expect(scope).not.toBeNull();
     expect(scope).toBeInstanceOf(WireScope);