+++ src/wirestate-react/test-utils.ts
@@ -1,7 +1,0 @@
-/**
- * React test helpers for rendering components with Wirestate container providers.
- *
- * @packageDocumentation
- */
-
-export { withContainerProvider } from "./test-utils/with-container-provider";