+++ src/wirestate/test-utils.ts
@@ -1,1 +1,0 @@
-export * from "../wirestate-react/test-utils";